backend/dev-requirements.txt
-r requirements.txt

black>=20.8b1
mypy>=0.790
pylint>=2.6.0
pylint-flask>=0.6
pylint-flask-sqlalchemy>=0.2.0

jsonpath-ng>=1.5.2
parameterized>=0.7.4
pytest>=6.2.1
pytest-cov>=2.10.1
pytest-flask>=1.1.0
coverage>=5.3.1
